🧹 back: move routing maritime utils to a dedicated file
